+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log (@egomobile/orm-pg)
 
+## 0.12.0
+
+- **BREAKING CHANGE**: update to version `^0.9.0` of [node-orm](https://github.com/egomobile/node-orm)
+- library required at least [Node 16](https://nodejs.org/en/blog/release/v16.0.0/) now
+- add `registerBigIntAsNumber()` function, which can help to fix issue, that `bigint` is returned as `string` instead of `number`
+- `npm update`s
+- add missing documentation
+
 ## 0.11.0
 
 - **BREAKING CHANGE**: (module property of IPostgreSQLMigration interface)[https://egomobile.github.io/node-orm-pg/interfaces/IPostgreSQLMigration.html#module] is now lazy loaded and read-only